Modified sensor and device command relative domain classes annotations.

+++ branches/Modeller_gwt_Branch/src/org/openremote/modeler/service/impl/SensorServiceImpl.java
@@ -3,27 +3,32 @@ package org.openremote.modeler.service.impl;
 import java.util.List;
 
 import org.hibernate.Hibernate;
+import org.hibernate.criterion.DetachedCriteria;
+import org.hibernate.criterion.Restrictions;
 import org.openremote.modeler.domain.Account;
+import org.openremote.modeler.domain.CommandRefItem;
 import org.openremote.modeler.domain.CustomSensor;
 import org.openremote.modeler.domain.Sensor;
+import org.openremote.modeler.domain.SensorRefItem;
 import org.openremote.modeler.service.BaseAbstractService;
 import org.openremote.modeler.service.SensorService;
 
 public class SensorServiceImpl extends BaseAbstractService<Sensor> implements SensorService {
 
-   public void deleteSensor(long id) {
+   public Boolean deleteSensor(long id) {
       Sensor sensor = super.loadById(id);
-            genericDAO.delete(sensor);
+      DetachedCriteria criteria = DetachedCriteria.forClass(SensorRefItem.class);
+      List<SensorRefItem> sensorRefItems = genericDAO.findByDetachedCriteria(criteria.add(Restrictions.eq("sensor", sensor)));
+      if (sensorRefItems.size() > 0) {
+         return false;
+      } else {
+         genericDAO.delete(sensor);
+      }
+      return true;
    }
 
    public List<Sensor> loadAll(Account account) {
-      List<Sensor> sensors = account.getSensors();
-      for (Sensor sensor : sensors) {
-         if(sensor instanceof CustomSensor) {
-            Hibernate.initialize(((CustomSensor)sensor).getStates());
-         }
-      }
-      return sensors;
+      return account.getSensors();
    }
 
    public Sensor saveSensor(Sensor sensor) {
@@ -40,7 +45,11 @@ public class SensorServiceImpl extends BaseAbstractService<Sensor> implements Se
    }
 
    public Sensor loadById(long id) {
-      return genericDAO.getById(Sensor.class, id);
+      Sensor sensor = genericDAO.getById(Sensor.class, id);
+      if(sensor instanceof CustomSensor) {
+         Hibernate.initialize(((CustomSensor)sensor).getStates());
+      }
+      return sensor;
    }
 
 }
